+++ pymongo/encryption_options.py
@@ -70,7 +70,7 @@ class AutoEncryptionOpts:
         users. To configure automatic *decryption* without automatic
         *encryption* set ``bypass_auto_encryption=True``. Explicit
         encryption and explicit decryption is also supported for all users
-        with the :class:`~pymongo.encryption.ClientEncryption` class.
+        with the :class:`~pymongo.asynchronous.encryption.AsyncClientEncryption` and :class:`~pymongo.encryption.ClientEncryption` classes.
 
         See :ref:`automatic-client-side-encryption` for an example.
 
@@ -238,8 +238,8 @@ class RangeOpts:
 
     def __init__(
         self,
-        sparsity: int,
-        trim_factor: int,
+        sparsity: Optional[int] = None,
+        trim_factor: Optional[int] = None,
         min: Optional[Any] = None,
         max: Optional[Any] = None,
         precision: Optional[int] = None,
@@ -264,7 +264,7 @@ class RangeOpts:
     def document(self) -> dict[str, Any]:
         doc = {}
         for k, v in [
-            ("sparsity", int64.Int64(self.sparsity)),
+            ("sparsity", int64.Int64(self.sparsity) if self.sparsity else None),
             ("trimFactor", self.trim_factor),
             ("precision", self.precision),
             ("min", self.min),